Few more refactoring classes
