+++ app/src/demo/java/com.karthik.splash.root/SplashNetworkModule.kt
@@ -2,10 +2,10 @@ package com.karthik.splash.root
 
 import android.content.Context
 import com.chuckerteam.chucker.api.ChuckerInterceptor
+import com.karthik.network.utility.AuthorizationKeyInterceptor
+import com.karthik.network.utility.UserOfflineInterceptor
 import com.karthik.splash.BuildConfig
 import com.karthik.splash.misc.InternetHandler
-import com.karthik.splash.restserviceutility.interceptors.AuthorizationKeyInterceptor
-import com.karthik.splash.restserviceutility.interceptors.UserOfflineInterceptor
 import com.karthik.splash.storage.MemoryCache
 import dagger.Module
 import dagger.Provides
