+++ app/src/main/java/com/karthik/splash/misc/Utils.kt
@@ -3,13 +3,11 @@ package com.karthik.splash.misc
 import com.google.gson.Gson
 import com.google.gson.reflect.TypeToken
 import com.karthik.network.home.bottomliketab.models.Photos
-import com.karthik.splash.restserviceutility.UserOfflineException
 import java.text.ParseException
 import java.text.SimpleDateFormat
 import java.util.*
 
 class Utils {
-    enum class NetworkErrorType { OFFLINE, IOEXCEPTION }
     enum class ResponseType { NEW, TRENDING, FEATURED, LIKES }
 
     companion object {
@@ -21,25 +19,6 @@ class Utils {
         const val imageLoadDuration = 1200
         const val pageSize = 10
         const val intialPageSize = 10
-
-        fun convertArrayListToString(responseList: List<Any>): String =
-                Gson().toJson(responseList)
-
-        fun convertStringToArrayList(
-                response: String,
-                responseType: ResponseType
-        ): ArrayList<out Any>? {
-            if (responseType == ResponseType.NEW || responseType == ResponseType.TRENDING
-                    || responseType == ResponseType.FEATURED) {
-                val type = object : TypeToken<ArrayList<Photos>>() {}.type
-                return Gson().fromJson<ArrayList<out Any>>(response, type)
-            }
-            return null
-        }
-
-        fun getErrorType(e: Throwable): Utils.NetworkErrorType =
-                if (e is UserOfflineException) Utils.NetworkErrorType.OFFLINE
-                else Utils.NetworkErrorType.IOEXCEPTION
 
         fun parseDate(receivedDate: String?): String? {
             if (receivedDate == null)
