--- app/src/main/java/com/karthik/splash/restserviceutility/UserOfflineException.kt
@@ -1,10 +1,0 @@
-package com.karthik.splash.restserviceutility
-
-
-import java.io.IOException
-
-/**
- * Created by karthikrk on 19/11/17.
- */
-
-class UserOfflineException : IOException()
