+++ network/src/main/java/com/karthik/network/utility/UserOfflineInterceptor.kt
@@ -1,20 +1,17 @@
-package com.karthik.splash.restserviceutility.interceptors
+package com.karthik.network.utility
 
 
-import com.karthik.splash.misc.InternetHandler
-
-import com.karthik.splash.restserviceutility.UserOfflineException
-
-import java.io.IOException
-
+import com.karthik.network.IInternetHandler
+import com.karthik.network.UserOfflineException
 import okhttp3.Interceptor
 import okhttp3.Response
+import java.io.IOException
 
 /**
  * Created by karthikrk on 19/11/17.
  */
 
-class UserOfflineInterceptor(private val internetHandler: InternetHandler) : Interceptor {
+class UserOfflineInterceptor(private val internetHandler: IInternetHandler) : Interceptor {
 
     @Throws(IOException::class)
     override fun intercept(chain: Interceptor.Chain): Response {
